requirements.txt
pydub
flask
torch
openai-whisper
pyaudio